routeviewmap fits bounds
